refactor : PlayerStatsTab UI와 컴포넌트 경계를 줄이기 위해 플레이어 탭분리했습니다
