+++ src/1_Page/Championship/ui/DashBoard/ui/PlayerStatsTab/constant/sortColumns.ts
