+++ src/1_Page/Championship/ui/DashBoard/ui/PlayerStatsTab/model/type.d.ts
@@ -1,0 +1,10 @@
+type UseSearchHandlerReturn = {
+  searchTerm: string;
+  setSearchTerm: React.Dispatch<React.SetStateAction<string>>;
+  sortConfig: {
+    key: keyof PlayerStats;
+    direction: "asc" | "desc";
+  };
+  handleSort: (key: keyof PlayerStats) => void;
+  displayPlayerStats: PlayerStats[];
+};
