+++ src/1_Page/Championship/ui/DashBoard/ui/PlayerStatsTab/model/useSearchHandler.ts
@@ -1,32 +1,8 @@
 import React from "react";
 
 const useSearchHandler = (
-  initialPlayerStats: PlayerStats[]
+  playerStats: PlayerStats[]
 ): UseSearchHandlerReturn => {
-  // playerStats를 내부 상태로 관리
-  const [playerStats, setPlayerStats] = React.useState<PlayerStats[]>(
-    initialPlayerStats || []
-  );
-
-  // 초기값 설정
-  React.useEffect(() => {
-    setPlayerStats(initialPlayerStats || []);
-  }, [initialPlayerStats]);
-
-  // idx를 통해 특정 선수 정보만 수정하는 핸들러 함수
-  const handleUpdatePlayer = (
-    playerListIdx: number,
-    updatedStats: Partial<PlayerStats>
-  ) => {
-    setPlayerStats((prevStats) =>
-      prevStats.map((player) =>
-        player.player_list_idx === playerListIdx
-          ? { ...player, ...updatedStats }
-          : player
-      )
-    );
-  };
-
   // 기존 검색 및 정렬 로직은 유지
   const [searchTerm, setSearchTerm] = React.useState<string>("");
   const [sortConfig, setSortConfig] = React.useState<{
@@ -80,7 +56,6 @@ const useSearchHandler = (
     sortConfig,
     handleSort,
     displayPlayerStats,
-    handleUpdatePlayer, // 새로 추가된 핸들러
   };
 };
 export default useSearchHandler;
