+++ src/1_Page/Championship/ui/DashBoard/ui/PlayerStatsTab/type.d.ts
@@ -1,0 +1,3 @@
+type PlayerStatsTabProps = {
+  playerStats: PlayerStats[];
+};
