+++ src/1_Page/Championship/ui/DashBoard/ui/PlayerStatsTab/ui/PlayerRow/type.d.ts
